Rename directory: controlBarButtons -> controlBar
