+++ src/components/workspace/controlBar/clearButton.tsx
